FIX: TimeSpeed

+++ Supernatural City/Assets/_Game/Scripts/Managers/TimeManager.cs
@@ -33,7 +33,7 @@ public class TimeManager : MonoBehaviour
         if (ByButton) { PauseScreen.SetActive(false); }
     }
 
-    public void SpeedTime(float Times)
+    public void SpeedTime(int Times)
     {
         Time.timeScale = 1 * Times;
     }
